Реализованы ДТО и сервис под эндпоинты получения и изменения статусов…

+++ main-server/src/main/java/main/server/events/dto/NewEventDto.java
@@ -42,7 +42,7 @@ public class NewEventDto {
     Boolean paid;
 
     @PositiveOrZero
-    Integer participantsLimit;
+    Long participantsLimit;
 
     Boolean requestModeration;
